backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
alembic==1.12.0
redis==5.0.0
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.3.0
pydantic-settings==2.0.3
requests==2.31.0
python-socketio==5.9.0
websockets==11.0.3
